[hotfix] fix variable type for top_p (#5313)

+++ applications/Colossal-LLaMA-2/inference_example.py
@@ -68,7 +68,7 @@ if __name__ == "__main__":
     parser.add_argument("--do_sample", type=bool, default=True, help="Set whether or not to use sampling")
     parser.add_argument("--temperature", type=float, default=0.3, help="Set temperature value")
     parser.add_argument("--top_k", type=int, default=50, help="Set top_k value for top-k-filtering")
-    parser.add_argument("--top_p", type=int, default=0.95, help="Set top_p value for generation")
+    parser.add_argument("--top_p", type=float, default=0.95, help="Set top_p value for generation")
     parser.add_argument("--input_txt", type=str, default="明月松间照，", help="The prompt input to the model")
     parser.add_argument("--prompt_style", choices=["sft", "pretrained"], default="sft", help="The style of the prompt")
     args = parser.parse_args()
